Add support for outputting posed planes in IK component (optional)
Introduce a new feature to output posed planes in the
`InverseKinematicsComponent` class. Added a `_outputPosedPlanesParameter`
field to manage this functionality and updated the context menu with a
new "Output Posed Axis Planes" option.

Implemented the `MenuItemClickOutputPlanes` event handler to toggle the
posed planes output and added serialization/deserialization support for
this parameter in the `Write` and `Read` methods.

Created the `GetPosedPlanesDataTree` method to transform posed axis
planes into a structured data tree for output. Updated the `SolveInstance`
method to populate the posed planes output parameter when enabled.